tile grid and player functionalility
added grid system that can track the current tile the player is standing on allowing for tile effects to be applied such as slow down. Added player movement along with input manager for a centralised place to poll and receive input along with being able to add directional mapping to keys.