added helper methods to various classes
Added onPathEnd method to N[C class to allow for code to be executed when an NPC reaches the end of its path. Updated character class to accomadate for  the status effect manager class now allowing for status efffects to be applied across all characters in a more flexible manner with every character having a status effect manage that can take in a map that holds known status effects that can be applied to the character that are then managed and applied to the character via the status effect manager.