Added level builder and level manufacturer
Tests performed:
dynamic object movement test
based on fixed frame rate and varying frame rate
sprite generator testing associative containers such as the std::map that stores all data
associated with the various texture assets including testing the find algorithm for the associative map.
Tested the animation objects capacibilities when it comes to playing animation along with ensuring that the vector that contains the frame for the animation size matches the expected number of frames.