added menu system
added a menu system that allows for sub menus and actions to be bound to UI elements such as buttons via the use of method/function pointers. It follows a tree like design pattern where the root node is the main menu for the game which then stores sub menus that can be reached through the main menu.