Testing the builder design patter for level areas
These test include a unit test for the level builder itself testing its ability to configure different properties of an assigned/constructed level area within it, ensuring that the values of these properties are cofigured according to desired values defined within a test fixture, also ensuring that these properties can be configured separately and in a combined manner. There is also an included intergration test for both the level manufactuerer and the builder as the manufacturer makes use of a level builder to construct the final level area with the desired properties.