Adding score menu system
Score menu system that makes use of the std::sort algorithm to sort player scores that are loaded from a file and display them within a sub menu. This system also makes use of a functor to define how the scores should be sorted for the std::sort algorithm defining how the string created that stores the score should be interpreted in terms of how the score text should be ordered within the scores sub menu.

Also added comments to the main files that define the structure of the menu system.