Testing the score menu
Testing the readScores method of the score menu class which will read scores from a file and sort them using the std::sort algorihtm along with a custom predicate for the std::sort algorithm in the form of a functor, which will extract the scores from each of the score strings to be displayed in the score menu ui and compare the score values. Allowing the scores strings in the vector storing them to be sorted by the std::sort algorithm.